+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+ - Palette Editor: Color picker could not pick black (0,0,0) due to UI scaling bug.
 
 ## [1.11.3] - 2021-04-25
 ### Changed